Objects have position, eulerAngle, scale attributes which can be set in their DerivedUpdate() functions and which automatically set the model to the correct place

+++ PA6/include/planet.h
@@ -8,7 +8,7 @@
 class Planet : public BaseObject{
 public:
     Planet(BaseObject *parent_, float orbitRadius_, float orbitRate_, float rotateRate_);
-    void DerivedUpdate(double dt);
+    void DerivedUpdate(float dt);
 
     //Turning rate in rev/sec
 	float rotateRate;
